// src/shared/api/generated/models/surveyResultResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * 곧감 API
 * 곧감 서비스 API 명세서
 * OpenAPI spec version: v0.0.1
 */
import type { CategoryPenaltyResponse } from './categoryPenaltyResponse';
import type { CharacterResponse } from './characterResponse';
import type { SurveyResultResponseGender } from './surveyResultResponseGender';

export interface SurveyResultResponse {
  /** 결과 ID */
  resultId: number;
  /** 결과 공유 토큰 */
  shareToken: string;
  /** 이름 */
  name: string;
  /** 생년월일 */
  birthDate: string;
  /** 성별 */
  gender: SurveyResultResponseGender;
  /** 예상 사망 나이 */
  expectedLife: number;
  /** 오늘의 한마디 */
  todayMessage: string;
  /** 경고 메시지 */
  warningMessage: string;
  /** 캐릭터 얼굴 */
  character: CharacterResponse;
  /** 카테고리별 차감된 수명 */
  categoryPenalties: CategoryPenaltyResponse[];
  /** 특별 준수 사항 (1~3개) */
  specialRules: string[];
}